+++ R/vm_resource_config.R
@@ -1,7 +1,7 @@
 #' Resource configuration functions for a virtual machine deployment
 #'
 #' @param username For `user_config`, the name for the admin user account.
-#' @param sshkey For `user_config`, string containing an SSH public key. Can be the key itself, or the name of the public key file.
+#' @param sshkey For `user_config`, the SSH public key. This can be supplied in a number of ways: as a string with the key itself; the name of the public key file; or an `AzureRMR::az_resource` object pointing to an SSH public key resource (of type "Microsoft.Compute/sshPublicKeys"). See the examples below.
 #' @param password For `user_config`, the admin password. Supply either `sshkey` or `password`, but not both; also, note that Windows does not support SSH logins.
 #' @param size For `datadisk_config`, the size of the data disk in GB. St this to NULL for a disk that will be created from an image.
 #' @param name For `datadisk_config`, the disk name. Duplicate names will automatically be disambiguated prior to VM deployment.
@@ -11,18 +11,56 @@
 #' @param publisher,offer,sku,version For `image_config`, the details for a marketplace image.
 #' @param id For `image_config`, the resource ID for a disk to use as a custom image.
 #'
+#' @examples
+#' \dontrun{
+#'
+#' ## user_config: SSH public key resource in Azure
+#' # create the resource
+#' keyres <- rg$create_resource(type="Microsoft.Compute/sshPublicKeys", name="mysshkey")
+#'
+#' # generate the public and private keys
+#' keys <- keyres$do_operation("generateKeyPair", http_verb="POST")
+#' keyres$sync_fields()
+#'
+#' # save the private key (IMPORTANT)
+#' writeBin(keys$privateKey, "mysshkey.pem")
+#'
+#' # create a new VM using the public key resource for authentication
+#' # you can then login to the VM with ssh -i mysshkey.pem <username@vmaddress>
+#' rg$create_vm("myvm", user_config("username", sshkey=keyres), config="ubuntu_20.04")
+#'
+#'
+#' ## user_config: SSH public key as a file
+#' rg$create_vm("myvm", user_config("username", sshkey="mysshkey.pub"), config="ubuntu_20.04")
+#'
+#'
+#' ## user_config: SSH public key as a string (read from a file)
+#' pubkey <- readLines("mysshkey.pub")
+#' rg$create_vm("myvm", user_config("username", sshkey=pubkey), config="ubuntu_20.04")
+#'
+#' }
+#'
 #' @rdname vm_resource_config
 #' @export
 user_config <- function(username, sshkey=NULL, password=NULL)
 {
+    keyres <- is_resource(sshkey) && tolower(sshkey$type) == "microsoft.compute/sshpublickeys"
+    key <- is.character(sshkey) || keyres
     pwd <- is.character(password)
-    key <- is.character(sshkey)
+
     if(!pwd && !key)
         stop("Must supply either a login password or SSH key", call.=FALSE)
     if(pwd && key)
         stop("Supply either a login password or SSH key, but not both", call.=FALSE)
 
-    if(key && file.exists(sshkey))
+    if(keyres)
+    {
+        sshkey <- gsub("\r\n", "", sshkey$properties$publicKey)
+        if(is_empty(sshkey))
+            stop("Supplied public key resource is uninitialized, run generateKeyPair first and save the returned keys",
+                 call.=FALSE)
+    }
+    else if(key && file.exists(sshkey))
         sshkey <- readLines(sshkey)
 
     structure(list(user=username, key=sshkey, pwd=password), class="user_config")